requirements.txt
Flask==3.0.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
opencv-python==4.8.1.78
mtcnn==0.1.1
Pillow==10.0.1
numpy==1.24.3
scikit-learn==1.3.0
Werkzeug==3.0.1
